+++ basic/tests/functional/submitFlightPlan/SubmitFplCharterCest.php
@@ -4,6 +4,7 @@ namespace tests\functional\submitFlightPlan;
 
 use tests\fixtures\AuthAssignmentFixture;
 use tests\fixtures\CredentialTypeAircraftTypeFixture;
+use tests\fixtures\CredentialTypeAirportAircraftFixture;
 use tests\fixtures\CredentialTypeFixture;
 use tests\fixtures\PilotCredentialFixture;
 use tests\fixtures\SubmittedFlightPlanFixture;
@@ -14,11 +15,12 @@ class SubmitFplCharterCest
 {
     public function _fixtures(){
         return [
-            'authAssignment'             => AuthAssignmentFixture::class,
-            'submittedFlightPlan'         => SubmittedFlightPlanFixture::class,
-            'credentialType'             => CredentialTypeFixture::class,
-            'credentialTypeAircraftType' => CredentialTypeAircraftTypeFixture::class,
-            'pilotCredential'            => PilotCredentialFixture::class,
+            'authAssignment'              => AuthAssignmentFixture::class,
+            'submittedFlightPlan'          => SubmittedFlightPlanFixture::class,
+            'credentialType'              => CredentialTypeFixture::class,
+            'credentialTypeAircraftType'  => CredentialTypeAircraftTypeFixture::class,
+            'credentialTypeAirportAircraft' => CredentialTypeAirportAircraftFixture::class,
+            'pilotCredential'             => PilotCredentialFixture::class,
         ];
     }
 
@@ -449,6 +451,30 @@ class SubmitFplCharterCest
 
         $count = \app\models\SubmittedFlightPlan::find()->count();
         $I->assertEquals(6, $count);
+    }
+
+    // -------------------------------------------------------------------------
+    // Credential bypass prevention
+    // -------------------------------------------------------------------------
+
+    public function prepareFplCharterBlockedForPilotWithoutCredential(\FunctionalTester $I)
+    {
+        // Pilot 8 (no credentials) bypasses and tries B738 (aircraft 2) to LEMD → 403
+        $I->amLoggedInAs(8);
+        $I->amOnRoute('submitted-flight-plan/prepare-fpl-charter', ['arrival' => 'lemd', 'aircraft_id' => '2']);
+
+        $I->seeResponseCodeIs(403);
+        $I->dontSee('Flight Plan Submission');
+    }
+
+    public function prepareFplCharterBlockedAtGclpWithoutMnps(\FunctionalTester $I)
+    {
+        // Pilot 7 (B738 Rating, no MNPS) bypasses and tries B738 to GCLP → 403
+        $I->amLoggedInAs(7);
+        $I->amOnRoute('submitted-flight-plan/prepare-fpl-charter', ['arrival' => 'gclp', 'aircraft_id' => '2']);
+
+        $I->seeResponseCodeIs(403);
+        $I->dontSee('Flight Plan Submission');
     }
 
     public function openPrepareFplCharterValidVFRToIFRPlan(\FunctionalTester $I)
