+++ basic/tests/functional/submitFlightPlan/SubmitFplTourCest.php
@@ -4,6 +4,7 @@ namespace tests\functional\submitFlightPlan;
 
 use tests\fixtures\AuthAssignmentFixture;
 use tests\fixtures\CredentialTypeAircraftTypeFixture;
+use tests\fixtures\CredentialTypeAirportAircraftFixture;
 use tests\fixtures\CredentialTypeFixture;
 use tests\fixtures\PilotCredentialFixture;
 use tests\fixtures\SubmittedFlightPlanFixture;
@@ -14,12 +15,13 @@ class SubmitFplTourCest
 {
     public function _fixtures(){
         return [
-            'authAssignment'             => AuthAssignmentFixture::class,
-            'submittedFlightPlan'         => SubmittedFlightPlanFixture::class,
-            'tourStage'                  => TourStageFixture::class,
-            'credentialType'             => CredentialTypeFixture::class,
-            'credentialTypeAircraftType' => CredentialTypeAircraftTypeFixture::class,
-            'pilotCredential'            => PilotCredentialFixture::class,
+            'authAssignment'              => AuthAssignmentFixture::class,
+            'submittedFlightPlan'          => SubmittedFlightPlanFixture::class,
+            'tourStage'                   => TourStageFixture::class,
+            'credentialType'              => CredentialTypeFixture::class,
+            'credentialTypeAircraftType'  => CredentialTypeAircraftTypeFixture::class,
+            'credentialTypeAirportAircraft' => CredentialTypeAirportAircraftFixture::class,
+            'pilotCredential'             => PilotCredentialFixture::class,
         ];
     }
 
@@ -450,6 +452,30 @@ class SubmitFplTourCest
 
         $count = \app\models\SubmittedFlightPlan::find()->count();
         $I->assertEquals(6, $count);
+    }
+
+    // -------------------------------------------------------------------------
+    // Credential bypass prevention
+    // -------------------------------------------------------------------------
+
+    public function prepareFplTourBlockedForPilotWithoutCredential(\FunctionalTester $I)
+    {
+        // Pilot 8 (no credentials) bypasses and tries B738 (aircraft 2) on stage 2 (active LEBL→LEMD) → 403
+        $I->amLoggedInAs(8);
+        $I->amOnRoute('submitted-flight-plan/prepare-fpl-tour', ['tour_stage_id' => '2', 'aircraft_id' => '2']);
+
+        $I->seeResponseCodeIs(403);
+        $I->dontSee('Flight Plan Submission');
+    }
+
+    public function prepareFplTourBlockedAtGclpWithoutMnps(\FunctionalTester $I)
+    {
+        // Pilot 7 (B738 Rating, no MNPS) bypasses and tries B738 to GCLP (stage 3) → 403
+        $I->amLoggedInAs(7);
+        $I->amOnRoute('submitted-flight-plan/prepare-fpl-tour', ['tour_stage_id' => '3', 'aircraft_id' => '2']);
+
+        $I->seeResponseCodeIs(403);
+        $I->dontSee('Flight Plan Submission');
     }
 
     public function openPrepareFplTourValidVFRToIFRPlan(\FunctionalTester $I)
